threshold THRESH_BINARY
A binary threshold is a simple "either or" threshold, where the pixels
are either 255 or 0. In many cases, this would be white or black, but we
have left our image colored for now, so it may be colored still. The
first parameter here is the image. The next parameter is the threshold,
we are choosing 10. The next is the maximum value, which we're choosing
as 255. Next and finally we have the type of threshold, which we've
chosen as THRESH_BINARY. Normally, a threshold of 10 would be somewhat
poor of a choice. We are choosing 10, because this is a low-light
picture, so we choose a low number. Normally something about 125-150
would probably work best.